requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
trafilatura>=2.0.0
lxml>=4.9.0
google-generativeai>=0.3.0
chromadb>=0.4.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0
urllib3<2.3.0
numpy>=1.24.0
matplotlib>=3.5.0
pysqlite3>=2.6.0
streamlit-extras>=0.2.0